Enhance movie booking system functionality and UI
- Updated `MovieController.cs` to add `GetSeatsByShowtime` method and new using directives.
- Changed `Quantity` type in `TicketCombo.cs` and `SeatNumber` type in `Seat.cs` from `short?` to `int?`.
- Refactored `Details.cshtml` to improve layout, remove commented code, and update image paths.
- Modified `_ShowtimesPartial.cshtml` for better showtime item display and JavaScript handling.
- Refactored `details.js` for improved cinema item toggling and dynamic loading of showtime items.
- Enhanced seat selection logic in `seats.js` and added asynchronous fetching of seat data.
- Overall improvements to data handling, UI responsiveness, and code maintainability.

+++ DoAn/Models/Cinema/Seat.cs
@@ -14,7 +14,7 @@ public partial class Seat
 
     public string? SeatRow { get; set; }
 
-    public short? SeatNumber { get; set; }
+    public int? SeatNumber { get; set; }
 
     public virtual Room Room { get; set; } = null!;
